Rename functions.py to config.py
